requirements.txt
pillow
numpy
rembg[gpu]